validations for saving and updating - score of expert is transient (avg of ratings)

+++ src/main/java/com/freshome/dto/expert/ExpertResponseDTO.java
@@ -15,7 +15,7 @@ public record ExpertResponseDTO(
         UserStatus status,
         String phoneNumber,
         byte[] profileImage,
-        Long score
+        Double score
 //        Long creditId
 ) {
 }
